Module ovrt_sys::api

Note: At this time any API call that returns a value needs a callback assigned to return the value to, the 'callback' value will just be the string name of a function.

Any API call that returns a value also had an additional string optional parameter, this can be used to share data back to the callback function.

Modules

bindings
Raw bindings into the javascript API.

callbacks

Functions

broadcast_message
Send message all other open browser instances. (Calls 'ReceiveMessage').

close_overlay
Close the specified overlay.

get_finger_curls
Get finger curl positions.

get_monitor_count
(Used for SetContents monitorId).

get_overlay_bounds
Get bounds of overlay bounding box. (Refer to Unity documentation 'Bounds' section).

get_overlay_transform
Get OVROverlayTransform of a specified overlay.

get_overlay_transform_with_callback
Get OVROverlayTransform of a specified overlay.

get_overlay_type
Get type of overlay. (Browser, window capture, desktop capture).

get_window_titles
Returns a list of open windows and their handles. (If user has this option enabled).

refresh
Refresh a browser page.

send_device_data
Send device position/rotation data to the calling overlay. (Refer to 'Events' below).

send_message
Send message to specific browser instance. (Calls 'ReceiveMessage').

send_overlay_positions
Send overlay position/rotation data to the calling overlay. (Refer to 'Events' below).

set_browser_title
Sets the title of the browser that is visible in the Window List of OVR Toolkit.

set_contents_desktop
Set contents of an overlay.

set_contents_website
Set contents of an overlay.

set_contents_window
Set contents of an overlay.

set_keyboard_focus
Set if this overlay should receive keyboard inputs. (This will block the keyboard working on OVR Toolkit windows!).

set_overlay_position
Set position of an overlay.

set_overlay_rotation
Set rotation of an overlay. (EulerAngles).

set_overlay_setting_bool
Set overlay setting.

set_overlay_setting_f64
Set overlay setting.

set_overlay_setting_i32
Set overlay setting.

spawn_overlay
Spawn a new overlay.

spawn_overlay_with_callback
Spawn a new overlay.

submit