Module ovrt_sys::cmd

Modules

sink
CommandSink static instance and related types and functions.

Enums

Callback
Callback are pre-registered functions by ovrt-sys in which ovrt calls after some job/query has finished.

Command

Event

EventResponse

Notification
Notifications only exists in ovrt-sys and may be used to better control the app. Everytime a request is made for ovrt, a Notification is also sent by ovrt-sys, feedbacking itself.