Module ovrt_sys::cmd::sink

CommandSink static instance and related types and functions.

ovrt-sys must register some functions that are called by ovrt. Those functions then must forward some events/callback calls using the static instance of CommandSink, since there are no other parameters for the events/callbacks.

An UI user, such as an app using the druid framework, must then register itself as the listener of CommandSink, by being the first caller of global_command_sink and registering itself.

Structs

CommandSink

Functions

global_command_sink
This function has two use-cases: getting the CommandSink or initializing it.

submit_cmd
Submits a new command to be forwarded to the listener, or logged.

Type Definitions

Listener