Module ovrt_sys::events::bindings

Raw bindings into the javascript API.

For using those functions, prefer events instead.

Functions

__wasm_bindgen_generated_APIInit
Called when the API has finished injecting into the browser and the API can now be used.

__wasm_bindgen_generated_DevicePositionUpdate
Sends HMD and left/right controller position and rotation, also shows the active controller. (Needs to be enabled per overlay, refer to API above).

__wasm_bindgen_generated_InteractionStateChanged
If the user is interacting with the current overlay. (Mouse over).

__wasm_bindgen_generated_OverlayClosed
Called when an overlay is closed.

__wasm_bindgen_generated_OverlayOpened
Called when an overlay is spawned.

__wasm_bindgen_generated_OverlayTransformChanged
Called when an overlay is moved or its size changes. (Needs to be enabled per overlay, refer to API above).

__wasm_bindgen_generated_ReceiveMessage
Receives messages from other browser instances.

api_init
Called when the API has finished injecting into the browser and the API can now be used.

device_position_update
Sends HMD and left/right controller position and rotation, also shows the active controller. (Needs to be enabled per overlay, refer to API above).

interaction_state_changed
If the user is interacting with the current overlay. (Mouse over).

overlay_closed
Called when an overlay is closed.

overlay_opened
Called when an overlay is spawned.

overlay_transform_changed
Called when an overlay is moved or its size changes. (Needs to be enabled per overlay, refer to API above).

receive_message
Receives messages from other browser instances.