Function ovrt_sys::events::bindings::device_position_update

pub fn device_position_update(device_info: String)

Sends HMD and left/right controller position and rotation, also shows the active controller. (Needs to be enabled per overlay, refer to API above).